move top & bottom nav Appearance config to separate file
